Add files via upload
As someone who used to game, specifically playing Valorant frequently, I would aim train almost every day to keep my reflexes fast. One of the practice types you can do on the site is a line of dots you click on, or "shoot", as fast as you can. Your score is determined by how accurate your clicks are on the circles and how fast you're clicking each one. When I saw a way to make a wave with circles I immediately thought of an aim trainer and wanted to make one the same speed as what I used to use.
I was really struggling because I wanted to add an ellipse array as the background for my wave, and realized you can't stack them easily with the tools I have, but I spoke to my dad and he is going to get on a call and see if he can help me figure it out. It sounds silly, but the main thing I learned this week was that I have someone who can help me sometimes since he knows Python and Linux and uses it for work.